+++ packages-integrations/svelte-scoped/src/_common/transformApply/getUtils.test.ts
