+++ packages-integrations/svelte-scoped/src/_common/transformApply/index.test.ts
