+++ packages-integrations/svelte-scoped/src/_common/transformApply/index.ts
