+++ packages-integrations/svelte-scoped/src/_common/transformApply/removeOuterQuotes.test.ts
