+++ packages-integrations/svelte-scoped/src/_common/transformApply/writeUtilStyles.ts
