+++ packages-integrations/svelte-scoped/src/_common/transformClasses/addGeneratedStyles.ts
