+++ packages-integrations/svelte-scoped/src/_common/transformClasses/findClasses.ts
