+++ packages-integrations/svelte-scoped/src/_common/transformClasses/generateClassName.test.ts
