+++ packages-integrations/svelte-scoped/src/_common/transformClasses/hash.ts
