+++ packages-integrations/svelte-scoped/src/_common/transformClasses/index.test.ts
