+++ packages-integrations/svelte-scoped/src/_common/transformClasses/isShortcut.ts
