+++ packages-integrations/svelte-scoped/src/_common/transformClasses/needsGenerated.ts
