+++ packages-integrations/svelte-scoped/src/_common/transformClasses/parse.bench.ts
