+++ packages-integrations/svelte-scoped/src/_common/transformClasses/processClassBody.test.ts
