+++ packages-integrations/svelte-scoped/src/_common/transformClasses/processClasses.ts
