+++ packages-integrations/svelte-scoped/src/_common/transformClasses/processDirective.ts
