+++ packages-integrations/svelte-scoped/src/_common/transformClasses/processExpressions.test.ts
