+++ packages-integrations/svelte-scoped/src/_common/transformClasses/sortClassesIntoCategories.ts
