+++ packages-integrations/svelte-scoped/src/_common/transformClasses/test-outputs/EverythingDev.svelte
