+++ packages-integrations/svelte-scoped/src/_common/transformTheme.ts
