+++ packages-integrations/svelte-scoped/src/_preprocess/wrapGlobal.test.ts
